feat: add editing cabin

+++ src/services/apiCabins.ts
@@ -20,27 +20,46 @@ export async function deleteCabin(id: string) {
   return data;
 }
 
-export async function createCabin(newCabin: CreateCabin) {
-  const image = newCabin.image[0];
-  const imageName = `${Math.random()}-${image.name}`.replaceAll('/', '');
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images//${imageName}`;
-  const { data, error } = await supabase
-    .from('cabins')
-    .insert([{ ...newCabin, image: imagePath }])
-    .select();
-  if (error) {
-    throw new Error('Cabin could not be deleted');
-  }
-  const { error: storageError } = await supabase.storage
-    .from('cabin-images')
-    .upload(imageName, image, {
-      cacheControl: '3600',
-      upsert: false,
-    });
+export async function createEditCabin(newCabin: CreateCabin, id?: string) {
+  let imagePath = newCabin.image;
+  let imageName = '';
 
-  if (storageError) {
-    await supabase.from('cabins').delete().eq('id', data.id);
-    throw new Error('Cabin image could not be uploaded. Cabin was not created');
+  if (newCabin.image && typeof newCabin.image !== 'string') {
+    const image = newCabin.image?.[0];
+    imageName = `${Math.random()}-${image?.name}`.replaceAll('/', '');
+    imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images//${imageName}`;
+  }
+
+  let query = await supabase.from('cabins');
+
+  if (!id) {
+    query = query.insert([{ ...newCabin, image: imagePath }]);
+  } else {
+    query = query.update({ ...newCabin, image: imagePath }).eq('id', id);
+  }
+
+  const { data, error } = await query.select();
+
+  if (error) {
+    console.log(error);
+
+    throw new Error('Cabin could not be created');
+  }
+
+  if (typeof newCabin.image !== 'string') {
+    const { error: storageError } = await supabase.storage
+      .from('cabin-images')
+      .upload(imageName, newCabin.image?.[0], {
+        cacheControl: '3600',
+        upsert: false,
+      });
+
+    if (storageError) {
+      await supabase.from('cabins').delete().eq('id', data.id);
+      throw new Error(
+        'Cabin image could not be uploaded. Cabin was not created'
+      );
+    }
   }
 
   return data;
